tests/wpt/tests/svg/path/parsing/path-length-valid.tentative.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml">
  <title>SVG Path Properties: parsing path-length with valid values</title>
  <metadata>
    <h:link rel="help" href="https://w3c.github.io/svgwg/svg2-draft/paths.html#PathLengthProperty"/>
    <h:meta name="assert" content="path-length supports the full grammar 'none | &lt;length [0,∞]&gt;'."/>
  </metadata>
  <path id="target"/>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("path-length", "none");
test_valid_value("path-length", "0", "0px");
test_valid_value("path-length", "100px");
test_valid_value("path-length", "10em");
test_valid_value("path-length", "2rem");
test_valid_value("path-length", "calc(1px + 2px)", "calc(3px)");
test_valid_value("path-length", "calc(10px + 2em)", "calc(2em + 10px)");

  ]]></script>
</svg>
